requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
PyYAML==6.0.1
python-dateutil==2.8.2
lxml==5.1.0
pytest==7.4.3
